requirements.txt
fastapi==0.109.2
python-dotenv==1.0.0
motor==3.3.2
gunicorn==21.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
MetaTrader5==5.0.45